+++ src/components/GuestRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../context/AuthContext';
+import { useAuth } from '../context/useAuth';
 
 /**
  * Wraps guest-only routes — redirects to /dashboard when there is an active session.
