+++ src/components/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../context/AuthContext';
+import { useAuth } from '../context/useAuth';
 
 /**
  * Wraps protected routes — redirects to /signin when there is no active session.
