+++ src/context/DataContext.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const DataContext = React.createContext();
