+++ src/context/useAuth.js
@@ -1,0 +1,13 @@
+import { createContext, useContext } from 'react';
+
+export const AuthContext = createContext(null);
+
+export function useAuth() {
+  const context = useContext(AuthContext);
+
+  if (!context) {
+    throw new Error('useAuth must be used within an AuthProvider');
+  }
+
+  return context;
+}
