+++ src/pages/ResetPassword.jsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { supabase } from '../lib/supabaseClient';
-import { useAuth } from '../context/AuthContext';
+import { useAuth } from '../context/useAuth';
 import { Eye, EyeOff, Lock, CheckCircle2, Loader2, AlertTriangle } from 'lucide-react';
 
 export default function ResetPassword() {
